+++ tests/test_header.py
@@ -1,6 +1,6 @@
 from textual.app import App
 from textual.screen import Screen
-from textual.widgets import Header
+from textual.widgets import Header, Static
 
 
 async def test_screen_title_none_is_ignored():
@@ -16,7 +16,7 @@ async def test_screen_title_none_is_ignored():
 
     app = MyApp()
     async with app.run_test():
-        assert app.screen.query_one("HeaderTitle").text == "app title"
+        assert app.screen.query_one("HeaderTitle", Static).content == "app title"
 
 
 async def test_screen_title_overrides_app_title():
@@ -34,7 +34,7 @@ async def test_screen_title_overrides_app_title():
 
     app = MyApp()
     async with app.run_test():
-        assert app.screen.query_one("HeaderTitle").text == "screen title"
+        assert app.screen.query_one("HeaderTitle", Static).content == "screen title"
 
 
 async def test_screen_title_reactive_updates_title():
@@ -54,7 +54,7 @@ async def test_screen_title_reactive_updates_title():
     async with app.run_test() as pilot:
         app.screen.title = "new screen title"
         await pilot.pause()
-        assert app.screen.query_one("HeaderTitle").text == "new screen title"
+        assert app.screen.query_one("HeaderTitle", Static).content == "new screen title"
 
 
 async def test_app_title_reactive_does_not_update_title_when_screen_title_is_set():
@@ -74,7 +74,7 @@ async def test_app_title_reactive_does_not_update_title_when_screen_title_is_set
     async with app.run_test() as pilot:
         app.title = "new app title"
         await pilot.pause()
-        assert app.screen.query_one("HeaderTitle").text == "screen title"
+        assert app.screen.query_one("HeaderTitle", Static).content == "screen title"
 
 
 async def test_screen_sub_title_none_is_ignored():
@@ -90,7 +90,10 @@ async def test_screen_sub_title_none_is_ignored():
 
     app = MyApp()
     async with app.run_test():
-        assert app.screen.query_one("HeaderTitle").sub_text == "app sub-title"
+        assert (
+            app.screen.query_one("HeaderTitle", Static).content
+            == "MyApp — app sub-title"
+        )
 
 
 async def test_screen_sub_title_overrides_app_sub_title():
@@ -108,7 +111,10 @@ async def test_screen_sub_title_overrides_app_sub_title():
 
     app = MyApp()
     async with app.run_test():
-        assert app.screen.query_one("HeaderTitle").sub_text == "screen sub-title"
+        assert (
+            app.screen.query_one("HeaderTitle", Static).content
+            == "MyApp — screen sub-title"
+        )
 
 
 async def test_screen_sub_title_reactive_updates_sub_title():
@@ -128,24 +134,7 @@ async def test_screen_sub_title_reactive_updates_sub_title():
     async with app.run_test() as pilot:
         app.screen.sub_title = "new screen sub-title"
         await pilot.pause()
-        assert app.screen.query_one("HeaderTitle").sub_text == "new screen sub-title"
-
-
-async def test_app_sub_title_reactive_does_not_update_sub_title_when_screen_sub_title_is_set():
-    class MyScreen(Screen):
-        SUB_TITLE = "screen sub-title"
-
-        def compose(self):
-            yield Header()
-
-    class MyApp(App):
-        SUB_TITLE = "app sub-title"
-
-        def on_mount(self):
-            self.push_screen(MyScreen())
-
-    app = MyApp()
-    async with app.run_test() as pilot:
-        app.sub_title = "new app sub-title"
-        await pilot.pause()
-        assert app.screen.query_one("HeaderTitle").sub_text == "screen sub-title"
+        assert (
+            app.screen.query_one("HeaderTitle", Static).content
+            == "MyApp — new screen sub-title"
+        )